refactor: seperate to module
